+++ Source/Client/Features/Transactions/Components/TransactionComponent.razor
@@ -1,0 +1,16 @@
+﻿@inherits TransactionComponentBase
+
+<div id="transaction" @onclick="OnClicked" class="bg-secondary p-2 m-2">
+  <h3>Transaction Date:</h3> <h4>@TxDate</h4>
+  <h3>To Address:</h3> <h4>@ToAddress</h4>
+  <h3>Transaction Amount:</h3> <h4>@Amount</h4>
+</div>
+
+@if (IsClicked == true)
+{
+  <style>
+    #transaction {
+      background-color: blue;
+    }
+  </style>
+}
