+++ Source/Server/Services/Transactions/GetTransactionsHandler.cs
@@ -38,8 +38,9 @@
             DateTime.Now, // TX date,
             342 * (aIndex * 100), // Tx Amount,
             .0000000194887342, // Tx Gas,
-            "0xs7s0030384a08642" // To Address
-          )
+            "0xs7s0030384a08642", // To Address
+            aIndex // TxId
+            )
         )
       );
